test: move test utils
