+++ test/dereferencing.test.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import fs from 'fs/promises';
 import { describe, it, expect } from 'vitest';
-import { importFresh } from './utils';
+import { importFresh } from './test-utils';
 import { openapiToTsJsonSchema } from '../src';
 
 const fixtures = path.resolve(__dirname, 'fixtures');
