+++ test/index.test.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { existsSync } from 'fs';
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { importFresh } from './utils';
+import { importFresh } from './test-utils';
 import { openapiToTsJsonSchema } from '../src';
 
 const fixtures = path.resolve(__dirname, 'fixtures');
